requirements.txt
Django==4.2.2
django-cors-headers==4.1.0
django-mssql-backend==2.8.1
djangorestframework==3.14.0
mssql-django==1.3
drf_yasg==1.21.6
# pyodbc==4.0.39
requests==2.31.0
applicationinsights==0.11.10
# Whitenoise for serving static files
whitenoise==6.4.0
# Linux web server and workers
gunicorn==20.1.0
uvicorn==0.22.0

# Pandas and numpy for dataframe processing on user input data
polars==0.20.13
pyarrow==15.0.1

# # azure connection
# azure-storage-blob==12.17.0
# azure-identity==1.14.0
